+++ passkey-authenticator/src/authenticator/get_assertion.rs
@@ -149,6 +149,7 @@ where
                 name: "".into(),
             }),
             number_of_credentials: None,
+            unsigned_extension_outputs: None,
         })
     }
 }